Bicycle 階層構造の作成
